+++ packages/embed-wallet/src/EmbedWallet.ts
@@ -144,7 +144,6 @@ export class EmbedWallet {
     network,
     context,
     popupMode = 'modal',
-    skipOnboardingHelloPage,
     connectOptions = {},
     appId = this.options.appId,
     sessionNamespace = this.options.sessionNamespace,
@@ -154,7 +153,6 @@ export class EmbedWallet {
     this.connectOptions = connectOptions;
     this.defaultContext = createContext(context);
     this.defaultContext.app.appId ||= appId;
-    this.defaultContext.app.skipOnboardingHelloPage = skipOnboardingHelloPage;
 
     const { sessionId } = getAuthRedirectResult();