+++ playground/Wallet.tsx
@@ -21,11 +21,12 @@ export const Wallet = () => {
       connectOptions: {
         mode: 'modal',
       },
-      skipOnboardingHelloPage: true,
 
       context: {
         whiteLabel: {
           var1: 'value',
+          skipLoginIntro: true,
+          mainColor: '#000000',
         },
 
         app: {
@@ -112,6 +113,9 @@ export const Wallet = () => {
 
   const handleSetContext = useCallback(async () => {
     wallet.setContext({
+      whiteLabel: {
+        skipLoginIntro: true,
+      },
       banner: {
         thumbnailUrl: nftImageUrl,
         badgeUrl: logoUrl,